+++ libs/common/src/lib/charts/index.ts
@@ -1,5 +1,6 @@
 export * from './bar-chart/bar-chart.component';
 export * from './empty-chart/empty-chart.component';
 export * from './line-chart/line-chart.component';
+export * from './range-chart';
 export * from './chart-utils';
 export * from './charts.module';