+++ libs/common/src/lib/charts/range-chart/range.models.ts
@@ -1,0 +1,13 @@
+export interface RangeChartData {
+  category: string;
+  average: number;
+  min: number;
+  max: number;
+}
+
+export interface DatedRangeChartData {
+  timestamp: string;
+  average: number;
+  min: number;
+  max: number;
+}