+++ libs/common/src/lib/metrics/index.ts
@@ -1,0 +1,2 @@
+export * from './metrics-page.component';
+export * from './remi-metrics.service';